% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshot.R
\name{analyze_one_plate}
\alias{analyze_one_plate}
\title{analyze data from a plate}
\usage{
analyze_one_plate(plate, dirFigure = NA, models = drModels(return = TRUE,
  verbose = FALSE)[["recommendedModels"]], unit = NA, interpolation = FALSE,
  con = stdout(), info = "", returnData = FALSE, ...)
}
\arguments{
\item{plate}{a list of data from one plate}

\item{dirFigure}{where to save figure; default is no figure to save}

\item{models}{models to be fitted; default is drexplroer recModels}

\item{unit}{unit of drug concentration}

\item{con}{connection to store log information; default is the console terminal; User can pass a file connection so that log will be saved in a file}

\item{info}{information passed from analyze_plate_list() about the plate}

\item{...}{additional parameters passed to split_plate such as CL='CellLine', Agent='Agent1', Dose='AgentConc1', Response='CellCount', ControlName='Control1'}
}
\value{
a list of the fits
}
\description{
this wraps over fit_one_experiment
}

