% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshot.R
\name{analyze_plate_list}
\alias{analyze_plate_list}
\title{analyze a list of plate data}
\usage{
analyze_plate_list(plate_list, dirFigure = NA, core = NA, parallel = TRUE,
  models = drModels(return = TRUE, verbose = FALSE)[["recommendedModels"]],
  unit = NA, interpolation = FALSE, con = stdout(), ...)
}
\arguments{
\item{plate_list}{a list of plate data}

\item{dirFigure}{where to save figure; default is no figure to save}

\item{core}{number of cores to register for parallel computing}

\item{parallel}{logical}

\item{models}{models to be fitted; default is drexplroer recModels}

\item{unit}{unit of drug concentration}

\item{con}{connection to store log information; default is the console terminal; User can pass a file connection so that log will be saved in a file
Notice that con is only possible if not using parallel (foreach does not allow multiple cores to simultaneously writting on the same file, as connection cannot be exported to different node)}

\item{...}{additional parameters for colname tracking}
}
\value{
a list
}
\description{
this wraps over analyze_one_plate and use parallel computing within plate
paralell is performed per plate as an unit
}

