% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshot.R
\name{fit_one_experiment}
\alias{fit_one_experiment}
\title{fit model on one experiment and allows to plot a jpeg figure for the curves}
\usage{
fit_one_experiment(x, models = drModels(return = TRUE, verbose =
  FALSE)[["recommendedModels"]], dirFigure = NA, CL = "CellLine",
  Agent = "Agent1", Dose = "AgentConc1", Response = "CellCount",
  ControlName = "Control1", PlateID = "PlateID", PlateName = "PlateName",
  debug = FALSE, unit = NA, interpolation = FALSE, con = stdout(),
  info = "")
}
\arguments{
\item{x}{a data frame for data from one experiment}

\item{models}{models to be fitted; default is drexplroer recModels}

\item{dirFigure}{folder path to store the figures. If set to NA, no figures will be output. With the returned list, it is easy to generate the figures though.
the figures here are one drug/one cell line per file.}

\item{CL}{colname for cell line}

\item{Agent}{colname for drug name}

\item{Dose}{colname for dose}

\item{Response}{colname for response}

\item{ControlName}{name for the control data so that it will not be treated as a drug.}

\item{PlateID}{colname for PlateID}

\item{PlateName}{colname for PlateName, usually the run date}

\item{unit}{unit of drug concentration}

\item{interpolation}{interpolation passed to fitOneExp}

\item{con}{connection to store log information; default is the console terminal; User can pass a file connection so that log will be saved in a file}

\item{info}{information passed from analyze_plate_list() about the plate}
}
\value{
a list of the fits for different models
}
\description{
this use fitOneExp() from drexplorer to analyze a lung moonshot experiement unit so that plate ID are tracked
}

