% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshot.R
\name{parse_res_from_exp_fit}
\alias{parse_res_from_exp_fit}
\title{parse the result from fit_one_experiment}
\usage{
parse_res_from_exp_fit(fit, whichSE = "IC50")
}
\arguments{
\item{whichSE}{a vector of IC names for SE values, i.e. s.e. for IC50 by whichSE=c('IC50')}
}
\value{
a table or list of fit
}
\description{
the result is a list
}

