% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshot.R
\name{parse_res_from_plate_list}
\alias{parse_res_from_plate_list}
\title{parse the result from analyze_plate_list}
\usage{
parse_res_from_plate_list(ll, type = c("ICtable", "fitList"),
  whichSE = "IC50")
}
\arguments{
\item{ll}{list as returned by analyze_plate_list}

\item{type}{whether to extract IC table or a list of fit for plotting}

\item{whichSE}{a vector of IC names for SE values, i.e. s.e. for IC50 by whichSE=c('IC50')}
}
\value{
a table or list of fit
}
\description{
the result is a nested list; on the top level, we have resL and ICtab; resL is another list each element is the fit for one drug; ICtab is a table for 
the IC values for all experiments in a plate.
}

