% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshot.R
\name{pruneICtable}
\alias{pruneICtable}
\title{a wrapper of pruneICvec to prune a IC table}
\usage{
pruneICtable(tab, IC_names, minD_name, maxD_name)
}
\arguments{
\item{tab}{a table containing IC values and possibly other info}

\item{IC_names}{colnames for IC values}

\item{minD_name}{colname for minimum administered dose}

\item{maxD_name}{colname for maximum administered dose}
}
\value{
a pruned table
}
\description{
a wrapper of pruneICvec to prune a IC table
}

