% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshot.R
\name{pruneICvec}
\alias{pruneICvec}
\title{prune a vector of IC values according to minDose and maxDose}
\usage{
pruneICvec(vec, minD, maxD)
}
\arguments{
\item{vec}{a vector of IC values}

\item{minD}{minimum dose adminised}

\item{maxD}{maximum dose adminised}
}
\value{
a vector of the pruned IC values
}
\description{
This is useful if we want to truncate non-achievable IC values (i.e. -30, 30) with observed minimum and maximum dose
}

