% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{shadowPlot}
\alias{shadowPlot}
\title{shadow plot for multiple fits}
\usage{
shadowPlot(fitResList, indFocus = 1, h = c(0.3, 0.5), main = NA,
  cex.main = 1, type = c("plot", "line"), tag = "", style = "simple",
  xlab = NA, ylab = NA, ylim = NA, xlim = NA, lwd = 2, lty = 1,
  fgcol = NA, bgcol = 80, alpha = NA)
}
\arguments{
\item{fitResList}{a list of fit result from multiple experiments}

\item{indFocus}{which experiment to highlight and other are treated as background shadow}

\item{h}{horizontal line added to the figure, i.e. indicating IC50, IC70}

\item{main}{main}

\item{cex.main}{cex.main to adjust main title size}

\item{style}{either 'full' or 'simple' indicating if data points to be added.}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{lwd}{line width for the curves}

\item{fgcol}{col for focused line}

\item{bgcol}{col for background lines; can be a vector so that each line have different color}

\item{alpha}{alpha blending for background lines}
}
\description{
this plots multiple curves from a set of experiments
}
\details{
when passing a list with one fit object, this can be used to plot the best model for a experiment
}

