% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshot.R
\name{split_plate}
\alias{split_plate}
\title{split data from a plate into multiple experiment unit (one cell line one drug).}
\usage{
split_plate(dat, CL = "CellLine", Agent = "Agent1", Dose = "AgentConc1",
  Response = "CellCount", ControlName = "Control1", PlateID = "PlateID",
  PlateName = "PlateName", minReplicate = 3)
}
\arguments{
\item{dat}{a data frame}

\item{CL}{colname for cell line}

\item{Agent}{colname for drug name}

\item{Dose}{colname for dose}

\item{Response}{colname for response}

\item{ControlName}{name for the control data so that it will not be treated as a drug.}

\item{PlateID}{colname for PlateID. this is for tracking, not used for split}

\item{PlateName}{colname for PlateName, usually the run date. this is for tracking, not used for split}

\item{minReplicate}{if a subset does not have nrow exceeding this, throw an error}
}
\value{
a list of data frame after split with the same columns as dat
}
\description{
This split function tries to split data from a plate which consists multiple drugs for the same cell line. The control data is shared across
multiple drugs. It returns a list of experiement unit which is dose response data for one drug and one cell line including the control data.
}
\details{
This is written specific to the lung moonshot screening data where on one plate, only a single cell line is profiled with multiple drugs; The control
data is shared across different drugs. Currently it only splits on drug and append control data to each drug. 
We may later need to augment this function so that it splits on both cell line and drug.
}

